refactor(quotes): extract update mutation lifecycle slice
