+++ backend/app/features/quotes/mutation/__init__.py
@@ -1,0 +1,5 @@
+"""Quote mutation slice exports."""
+
+from app.features.quotes.mutation.service import QuoteMutationService
+
+__all__ = ["QuoteMutationService"]
